+++ Tests/KeyboardKitTests/UIKit/Buttons/UIKeyboardButton+GesturesTests.swift
@@ -34,11 +34,11 @@ class UIKeyboardButton_GesturesTests: QuickSpec {
                 button.addKeyboardGestures(in: vc)
                 let oldTap = button.gestureRecognizers!.first { $0 is UITapGestureRecognizer }
                 let oldPress = button.gestureRecognizers!.first { $0 is UILongPressGestureRecognizer }
-                let oldRepeat = button.gestureRecognizers!.first { $0 is RepeatingGestureRecognizer }
+                let oldRepeat = button.gestureRecognizers!.first { $0 is UIRepeatingGestureRecognizer }
                 button.addKeyboardGestures(in: vc)
                 let newTap = button.gestureRecognizers!.first { $0 is UITapGestureRecognizer }
                 let newPress = button.gestureRecognizers!.first { $0 is UILongPressGestureRecognizer }
-                let newRepeat = button.gestureRecognizers!.first { $0 is RepeatingGestureRecognizer }
+                let newRepeat = button.gestureRecognizers!.first { $0 is UIRepeatingGestureRecognizer }
                 expect(newTap).toNot(be(oldTap))
                 expect(newPress).toNot(be(oldPress))
                 expect(newRepeat).toNot(be(oldRepeat))
@@ -49,7 +49,7 @@ class UIKeyboardButton_GesturesTests: QuickSpec {
                 button.addKeyboardGestures(in: vc)
                 let tap = button.gestureRecognizers?.first { $0 is UITapGestureRecognizer }
                 let press = button.gestureRecognizers?.first { $0 is UILongPressGestureRecognizer }
-                let repeating = button.gestureRecognizers?.first { $0 is RepeatingGestureRecognizer }
+                let repeating = button.gestureRecognizers?.first { $0 is UIRepeatingGestureRecognizer }
                 expect(tap).to(beNil())
                 expect(press).to(beNil())
                 expect(repeating).to(beNil())
@@ -59,7 +59,7 @@ class UIKeyboardButton_GesturesTests: QuickSpec {
                 button.addKeyboardGestures(in: vc)
                 let tap = button.gestureRecognizers!.first { $0 is UITapGestureRecognizer }
                 let press = button.gestureRecognizers!.first { $0 is UILongPressGestureRecognizer }
-                let repeating = button.gestureRecognizers!.first { $0 is RepeatingGestureRecognizer }
+                let repeating = button.gestureRecognizers!.first { $0 is UIRepeatingGestureRecognizer }
                 expect(tap).toNot(beNil())
                 expect(press).toNot(beNil())
                 expect(repeating).toNot(beNil())